move plot_footprints to plot module
