+++ osmnx/footprints.py
@@ -1,16 +1,12 @@
 """Download and plot footprints from OpenStreetMap."""
 
 import geopandas as gpd
-import matplotlib.pyplot as plt
-from descartes import PolygonPatch
-from matplotlib.collections import PatchCollection
 from shapely.geometry import LineString
 from shapely.geometry import MultiPolygon
 from shapely.geometry import Polygon
 from shapely.ops import polygonize
 from . import boundaries
 from . import downloader
-from . import plot
 from . import projection
 from . import settings
 from . import utils
@@ -582,88 +578,3 @@ def footprints_from_place(place, footprint_type='building', retain_invalid=False
     return _create_footprints_gdf(polygon, retain_invalid=retain_invalid,
                                   footprint_type=footprint_type, timeout=timeout,
                                   memory=memory, custom_settings=custom_settings)
-
-
-
-def plot_footprints(gdf, fig=None, ax=None, figsize=None, color='#333333',
-                    bgcolor='w', set_bounds=True, bbox=None, save=False,
-                    show=True, close=False, filename='image',
-                    file_format='png', dpi=600):
-    """
-    Plot a GeoDataFrame of footprints.
-
-    Parameters
-    ----------
-    gdf : geopandas.GeoDataFrame
-        GeoDataFrame of footprints
-    fig : figure
-    ax : axis
-    figsize : tuple
-    color : string
-        the color of the footprints
-    bgcolor : string
-        the background color of the plot
-    set_bounds : bool
-        if True, set bounds from either passed-in bbox or the spatial extent of the gdf
-    bbox : tuple
-        if True and if set_bounds is True, set the display bounds to this bbox
-    save : bool
-        whether to save the figure to disk or not
-    show : bool
-        whether to display the figure or not
-    close : bool
-        close the figure (only if show equals False) to prevent display
-    filename : string
-        the name of the file to save
-    file_format : string
-        the format of the file to save (e.g., 'jpg', 'png', 'svg')
-    dpi : int
-        the resolution of the image file if saving
-
-    Returns
-    -------
-    fig, ax : tuple
-
-    """
-
-    if fig is None or ax is None:
-        fig, ax = plt.subplots(figsize=figsize, facecolor=bgcolor)
-        ax.set_facecolor(bgcolor)
-
-    # extract each polygon as a descartes patch, and add to a matplotlib patch
-    # collection
-    patches = []
-    for geometry in gdf['geometry']:
-        if isinstance(geometry, Polygon):
-            patches.append(PolygonPatch(geometry))
-        elif isinstance(geometry, MultiPolygon):
-            for subpolygon in geometry:  # if geometry is multipolygon, go through each constituent subpolygon
-                patches.append(PolygonPatch(subpolygon))
-    pc = PatchCollection(patches, facecolor=color, edgecolor=color, linewidth=0, alpha=1)
-    ax.add_collection(pc)
-
-    if set_bounds:
-        if bbox is None:
-            # set the figure bounds to the polygons' bounds
-            left, bottom, right, top = gdf.total_bounds
-        else:
-            top, bottom, right, left = bbox
-        ax.set_xlim((left, right))
-        ax.set_ylim((bottom, top))
-
-    # turn off the axis display set the margins to zero and point the ticks in
-    # so there's no space around the plot
-    ax.axis('off')
-    ax.margins(0)
-    ax.tick_params(which='both', direction='in')
-    fig.canvas.draw()
-
-    # make everything square
-    ax.set_aspect('equal')
-    fig.canvas.draw()
-
-    fig, ax = plot._save_and_show(fig=fig, ax=ax, save=save, show=show, close=close,
-                                  filename=filename, file_format=file_format, dpi=dpi,
-                                  axis_off=True)
-
-    return fig, ax
